Begin refactoring to build rule HTML dynamically
This is a big win! The first function to do this is
`it_exchange_membership_addon_build_post_restriction_rules()`. The set
of rules are now built dynamically without the need for constant type
evaluation.

Introduces:

- IT_Exchange_Membership_Rule_Factory
- IT_Exchange_Membership_AbstractContent_Rule

Additionally a number of methods are added to the Content_RuleInterface
to make them more powerful. Notably the `matches_post()` method to
examine whether the current rule is applicable to a given post. As well
as functions to add and retrieve delay rules.

The next step will be implementing the evaluate() methods and using
them during the is_post_restricted functions.