Add support for custom delay messages.
Also introduces the `{time_until_available}` and `{available_date}`
tags for rendering the delay message. The `%d` option is 'deprecated'.

The return type for `evaluate_drip()` has been changed to return an
array of the failed rule and tested subscription. Still not in love
with this, but currently necessary due to how the theme API is
architected.