Use an abstraction to store delay meta instead of directly by post ID.
This removes the `get_post_for_delay()` method from the Delay rule
interface. This tied delay rules to only be used in post contexts. Now
the delayable content rule is passed to the delay rule and used as a
conduit for storing this meta.

The following methods are introduced to achieve this:

- get_delay_meta()
- update_delay_meta()
- delete_delay_meta()

Additionally, the `it_exchange_membership_addon_build_drip_rules()`
function is deprecated.