Add interface for a user's membership access
Users can gain membership access through either a subscription, or just
a one-time purchase. To faciliate this, we introduce a
`IT_Exchange_User_MembershipInterface` for managing that interaction.

For now, we only have a `IT_Exchange_Membership_Subscription_Driver`,
but next is implementing a simple transaction driver.

By depending upon an abstraction, rather than a concretion, we'll be
able to easily substitute a custom DB table for memberships on the next
major version. That'll be a few versions from now, so this will serve
as a good stopgap.