Use requestor's data to build membership message. Disallow forever -> renewing upgrades. "Fix" tests.
We fudged the calculations by a day for the tests. Because we can't mock `time()` and we can't use
Carbon ( PHP 5.4+ ), there are inaccuracies. with how the date time diff is calculated due to months in the year, etc... Still trying to see if there is a better way to do this than "fudging" the math.