Implement ITE_Object pattern and ITE_Proratable.
Simplifies most of the prorate handling for REST in the recurring-payments
add-on. Dynamically declares ITE_Proratable_User_Membership depending
on whether recurring payments is active or not.

While RP is necessary for prorating to work, it isn't for memberships to function
so we don't want to fatal error on a missing interface.